Add Adobe CC deal + new "Creative" category
Not sure on the category heading here, but felt it made sense to split off from "Training". Also needs checking to see if the URL will direct to a UK specific page. As I'm in the UK, it displays in £, but ideally this link should show $ or whatever the currency is in the country the user is coming from.